+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 2.2.0
+
+- Add _externals_ option to skip traversing some imports
+
+## 2.1.0
+
+- Add cache for `resolve` – speed up sync resolve version 2x
+
 ## 2.0.0
 
 - breaking: add `async filterDependent` as default export